chore: 더미데이터 consts로 이동
